👷 [ci] Disabling CodeQL scans while no python files exist.
Eventually, this repository will have python files in need of scanning. However, those files don't currently exist and this build task fails as a result.

We are temporarily disabling the task (by forcing it to look at git branches that do not exist). We should undo the effects of this change (by removing the branch name prefix) once python files are introduced into the repository.