✨ [+feature] Calculate project version based on git changes
The [semantic version](https://semver.org/) for a package is generated based on changes in git.

By default, the patch number is incremented with each commit. Include "+ minor" (without the space) in the commit's message to increment the minor version number and "+ major" (without the space) to increment the major version.